backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
crewai==0.28.8
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
openai>=1.13.3
anthropic==0.7.7
google-search-results==2.4.2
python-multipart==0.0.6
websockets==12.0
